fix(iot-device): Fix issue where SDK didn't validate application properties correctly
Normally the SDK wouldn't validate something like this, but if a user tries to send an application property with an invalid character in it, the different protocols behave differently in unintuitive ways. MQTT will sever the connection with no error message while AMQPS and HTTPS will allow it.

I'm also taking this opportunity to fix a hack tied to URL encoding. URLEncoder doesn't handle encoding spaces correctly, but URIUtil from apache commons does